If the buffer timeout elapses while a prompt (password) is in progress then retry after the prompt is complete.
 There is likely a cleaner way to do this but our decode is on a
 different thread to the data reader so we need to effectively pause the
 reader while the decoder is blocked by user input.

 We should really use wait handles.

+++ bass_zipstream/bass_zipstream.cpp
@@ -56,7 +56,7 @@ extern "C" {
 				if (factor > 1) {
 					factor = 1;
 				}
-				QWORD position = ARCHIVE_GetEntryLength(handle) * factor;
+				QWORD position = (QWORD)((FLOAT)ARCHIVE_GetEntryLength(handle) * factor);
 				ARCHIVE_BufferEntry(position, handle);
 			}
 			return BASS_StreamCreateFileUser(STREAMFILE_BUFFER, flags, &procs, handle);
